% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model.R
\name{age_model}
\alias{age_model}
\title{Age Model Using Bacon}
\usage{
age_model(data)
}
\arguments{
\item{data}{A data frame containing the necessary columns for age modeling.}
}
\value{
A list containing the age model results.
}
\description{
This function performs age modeling using the Bacon algorithm.
}
