% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_heatmap.R
\name{get_heatmap_data}
\alias{get_heatmap_data}
\title{Get Heatmap Data}
\usage{
get_heatmap_data(all_data)
}
\arguments{
\item{all_data}{A list containing data frames from each Excel sheet.}
}
\value{
A data frame ready for generating a heatmap.
}
\description{
This function retrieves and prepares data for generating a heatmap.
}
