% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar.R
\name{get_mar_sheet}
\alias{get_mar_sheet}
\title{Get MAR Sheet from Loaded Data and Plot}
\usage{
get_mar_sheet(all_data, x_breaks = NULL, x_labels = NULL)
}
\arguments{
\item{all_data}{A list containing data frames from each Excel sheet.}

\item{x_breaks}{Breaks for the x-axis.}

\item{x_labels}{Labels for the x-axis corresponding to the breaks.}
}
\value{
The mar data frame.
}
\description{
This function retrieves the Excel sheet called "mar" from the all_data list
and plots the mass accumulation rate.
}
