% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_rate.R
\name{get_sed_rate_sheet}
\alias{get_sed_rate_sheet}
\title{Get Sed Rate Sheet from Loaded Data and Plot}
\usage{
get_sed_rate_sheet(all_data, x_breaks = NULL, x_labels = NULL)
}
\arguments{
\item{all_data}{A list containing data frames from each Excel sheet.}

\item{x_breaks}{Breaks for the x-axis.}

\item{x_labels}{Labels for the x-axis corresponding to the breaks.}
}
\value{
The sed_rate data frame.
}
\description{
This function retrieves the Excel sheet called "sed_rate" from the all_data list
and plots the sedimentation rate.
}
