% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipd.R
\name{import_lipd}
\alias{import_lipd}
\title{Import Data from LiPD File}
\usage{
import_lipd(file_path)
}
\arguments{
\item{file_path}{Path to the LiPD file.}
}
\value{
A data frame containing the imported data.
}
\description{
This function imports data from a LiPD (Linked Paleo Data) file.
}
