% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_paleo_data}
\alias{load_paleo_data}
\title{Load Paleo Data from Excel File}
\usage{
load_paleo_data(excel_file)
}
\arguments{
\item{excel_file}{Path to the Excel file.}
}
\value{
A list containing data frames from each sheet in the Excel file.
}
\description{
This function loads paleoclimate data from an Excel file.
}
