% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bd.R
\name{plot_bd_ens}
\alias{plot_bd_ens}
\title{Plot Bulk Density Ensemble}
\usage{
plot_bd_ens(map_bd)
}
\arguments{
\item{map_bd}{Data frame containing the bulk density ensemble data.}
}
\value{
A ggplot object representing the bulk density ensemble plot.
}
\description{
This function plots the bulk density ensemble.
}
