% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bsi.R
\name{plot_bsi_ens}
\alias{plot_bsi_ens}
\title{Plot Biogenic Silica Ensemble}
\usage{
plot_bsi_ens(map_bsi)
}
\arguments{
\item{map_bsi}{Data frame containing the BSi ensemble data.}
}
\value{
A ggplot object representing the BSi ensemble plot.
}
\description{
This function plots the BSi ensemble.
}
