% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loi.R
\name{plot_loi_ens}
\alias{plot_loi_ens}
\title{Plot LOI Ensemble}
\usage{
plot_loi_ens(map_loi)
}
\arguments{
\item{map_loi}{Data frame containing the LOI ensemble data.}
}
\value{
A ggplot object representing the LOI ensemble plot.
}
\description{
This function plots the LOI ensemble.
}
