% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ms.R
\name{plot_ms_ens}
\alias{plot_ms_ens}
\title{Plot Magnetic Susptibility Ensemble}
\usage{
plot_ms_ens(map_ms)
}
\arguments{
\item{map_ms}{Data frame containing the bulk density ensemble data.}
}
\value{
A ggplot object representing the bulk density ensemble plot.
}
\description{
This function plots the bulk density ensemble.
}
