% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rabd.R
\name{plot_rabd_ens}
\alias{plot_rabd_ens}
\title{Plot RABD Ensemble}
\usage{
plot_rabd_ens(map_rabd)
}
\arguments{
\item{map_rabd}{Data frame containing the RABD ensemble data.}
}
\value{
A ggplot object representing the RABD ensemble plot.
}
\description{
This function plots the RABD ensemble.
}
