% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_heatmap.R
\name{psa_heatmap}
\alias{psa_heatmap}
\title{Generate PSA Heatmap}
\usage{
psa_heatmap(melted_df)
}
\arguments{
\item{melted_df}{A data frame containing melted PSA data.}
}
\value{
A ggplot object representing the PSA heatmap.
}
\description{
This function generates a heatmap of PSA distribution.
}
