% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_model.R
\name{plot_age_model}
\alias{plot_age_model}
\title{Plot Age Model}
\usage{
plot_age_model(
  age_model,
  x_breaks = NULL,
  x_labels = NULL,
  y_breaks = NULL,
  y_labels = NULL
)
}
\arguments{
\item{age_model}{An object obtained from age modeling.}

\item{x_breaks}{Breaks for the x-axis.}

\item{x_labels}{Labels for the x-axis corresponding to the breaks.}

\item{y_breaks}{Breaks for the y-axis.}

\item{y_labels}{Labels for the y-axis corresponding to the breaks.}
}
\value{
A ggplot object representing the age model plot.
}
\description{
This function generates a plot of age model using plotChronEns from geoChronR.
}
