% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model.R, R/plot_age_model.R
\name{plot_age_model}
\alias{plot_age_model}
\title{Plot Age Model}
\usage{
plot_age_model(age_model)

plot_age_model(age_model)
}
\arguments{
\item{age_model}{An object obtained from age modeling.}
}
\value{
A ggplot object representing the age model plot.

A ggplot object representing the age model plot.
}
\description{
This function generates a plot of age model using plotChronEns from geoChronR.

This function generates a plot of age model using plotChronEns from geoChronR.
}
